Rename scrambler to XORKeyScrambler
